abstract out screen table
